Use consistent type for vreplication id (#12218)
* Use consistent type for vreplication id

The `_vt.vreplication` table has an `INT` type without `UNSIGNED`, so it
maps to the Go `int32` type to have the exact same consistent range.

This means that we should use `int32` consistently for everywhere that
we use the VReplication identifier.

We had int32, uint32 and int used pretty randomly for these which leads
to CodeQL flagging these and how this was found.

Instead this now converts everything to int32. It's also safe to change
the protobuf, since the protobuf wire format for int32, uint32, int64 &
uint64 is all compatible and actually uses the same variable int length
format.

The only exception would be if someone has a VReplication identifier
that would not fit in `int32`, but well, this PR shows that in that case
we already blow up in unexpected and unknown ways anyway. So it's better
to have that explicitly error.


* Add support for all integer types for bind variables

This ensures we cover all the Go int types to generate bind parameters
from.


* Regenerate protos for vtadmin


* Fix test for using ParseInt


---------

+++ go/vt/vtctl/workflow/stream_migrator.go
@@ -214,7 +214,7 @@ func (sm *StreamMigrator) readTabletStreams(ctx context.Context, ti *topo.Tablet
 	tabletStreams := make([]*VReplicationStream, 0, len(qr.Rows))
 
 	for _, row := range qr.Named().Rows {
-		id, err := row["id"].ToInt64()
+		id, err := row["id"].ToInt32()
 		if err != nil {
 			return nil, err
 		}
@@ -268,7 +268,7 @@ func (sm *StreamMigrator) readTabletStreams(ctx context.Context, ti *topo.Tablet
 		}
 
 		tabletStreams = append(tabletStreams, &VReplicationStream{
-			ID:                 uint32(id),
+			ID:                 id,
 			Workflow:           workflowName,
 			BinlogSource:       &bls,
 			Position:           pos,
@@ -478,7 +478,7 @@ func (sm *StreamMigrator) syncSourceStreams(ctx context.Context) (map[string]mys
 				}
 
 				sm.ts.Logger().Infof("Waiting for keyspace:shard: %v:%v, position %v", sm.ts.SourceKeyspaceName(), shard, pos)
-				if err := sm.ts.TabletManagerClient().VReplicationWaitForPos(ctx, primary.Tablet, int(vrs.ID), mysql.EncodePosition(pos)); err != nil {
+				if err := sm.ts.TabletManagerClient().VReplicationWaitForPos(ctx, primary.Tablet, vrs.ID, mysql.EncodePosition(pos)); err != nil {
 					allErrors.RecordError(err)
 					return
 				}
